Make bytes_read a gsize from a guint for 64-bit cleanliness.
2003-11-13  Joe Shaw  <joe@ximian.com>

	* rc-line-buf.c (rc_line_buf_cb): Make bytes_read a gsize from
	a guint for 64-bit cleanliness.

+++ src/rc-line-buf.c
@@ -194,7 +194,7 @@ rc_line_buf_cb (GIOChannel *source, GIOCondition condition,
                 gpointer data)
 {
     RCLineBuf *line_buf = (RCLineBuf *)data;
-    guint bytes_read;
+    gsize bytes_read;
     guint count;
     guint base = 0;
     gchar buf[BUF_SIZE + 1];
